% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HSMMdata.R
\name{HSMMdata}
\alias{HSMMdata}
\title{Sinlge-cell RNA-seq data for HSMM (Human Skeletal Muscle Myoblast)}
\format{A matrix}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52529}
}
\description{
The R object is a log2 transformed expression matrix (HSMMdata) with 518 genes and 271 cells. This dataset is a subset of original gene expression profiles, which is available in the HSMMSingleCell package.
}
\references{
Trapnell, C., Cacchiarelli, D., Grimsby, J., Pokharel, P., Li, S., Morse, M., ... & Rinn, J. L. (2014). The dynamics and regulators of cell fate decisions are revealed by pseudotemporal ordering of single cells. Nature biotechnology.
}

