% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patternGOanalysis.R
\name{patternGOanalysis}
\alias{patternGOanalysis}
\title{patternGOanalysis}
\usage{
patternGOanalysis(pattern, type = NULL, termnum = 20,
  identifier = "ENSEMBL", species = "Human")
}
\arguments{
\item{pattern}{The direct output of truetimepattern or pseudotimepattern function.}

\item{type}{Character value of specific pattern to perform the GO analysis. If NULL GO analysis will be performed for all patterns.}

\item{termnum}{Number of top GO terms to be displayed.}

\item{identifier}{The identifier of the genes. It should be one of the following: "Entrez", "GenBank", "Alias", "Ensembl", "Gene", "Symbol", "GeneName" and "UniGene"}

\item{species}{The species of the genes. Currently only "Human" and "Mouse" are supported}
}
\value{
A list where each element is a data.frame containing the results of GO analysis.
}
\description{
Performs GO analysis on genes for each pattern
}
\details{
This function is basically a wrap up of functions in the topGO package. It takes as input the output of truetimepattern or pseudotimepattern function.
For each pattern, the GO enrichment analysis is performed where input genes are genes with specific patterns and background genes are all genes in the
expression profile. Users should correctly select identifier and species, otherwise the function may breakdown.
}
\examples{
library(topGO)
data(HSMMdata)
pattern <- truetimepattern(HSMMdata,truetime,removeconstant=TRUE)
patternGOanalysis(pattern,termnum=20,identifier="ENSEMBL",species="Human")
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

