% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/patternsummary.R
\name{patternsummary}
\alias{patternsummary}
\title{patternsummary}
\usage{
patternsummary(pattern)
}
\arguments{
\item{pattern}{The direct output of truetimepattern or pseudotimepattern function.}
}
\value{
A data.frame object. First column: pattern; Second column: number of genes
}
\description{
Count table for pattern
}
\details{
This function generates a count table of number of genes with each pattern.
}
\examples{
data(HSMMdata)
pattern <- truetimepattern(HSMMdata,truetime,removeconstant=TRUE)
patternsummary(pattern)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

