% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudotimepattern.R
\name{pseudotimepattern}
\alias{pseudotimepattern}
\title{pseudotimepattern}
\usage{
pseudotimepattern(expr, pseudotime, simplify = T, removeconstant = F,
  plot = F, gap = 10)
}
\arguments{
\item{expr}{The matrix of gene expression profile.}

\item{pseudotime}{A character data.frame or matrix of pseudo-time. First column: Cell name; Second column: pseudo-time.}

\item{simplify}{Whether to simplify pattern so that same neiboring patterns will be reduced to one. For example "up_up_constant" will be simplied to "up_constant".}

\item{removeconstant}{Whether to remove all constant patterns. For example "up_up_constant" will be simplied to "up_up". This step will be performed before simplify.}

\item{plot}{Whether to generate plot for genes with transition points.}

\item{gap}{Number of first and last gap cells that will be excluded when considering transition points.}
}
\value{
A list. expr: original expression matrix; pseudotime: original pseudotime; pattern: a list containing results of different patterns. For single patterns, it is a named
vector where values are the p-values of the t-test of the simple linear regression slope coefficient. The vector is ordered according to the p-values. For transition patterns,
a data.frame containing the mean and confidence interval of the transition point. It is ordered according to the transition points; fitexpr: the fitted expression matrix
}
\description{
Identify Pattern for Pseudo Temporal Cell Ordering
}
\details{
Identify the gene expression patterns for true experiment time. For the expressions of each gene, the function performs t-tests for cells from neighboring time points.
The expression pattern for cells from neiboring time points could be increasing, decreasing or constant. All patterns are concatenated using "-" to form the final pattern.
}
\examples{
data(HSMMdata)
pseudotimepattern(HSMMdata,pseudotime)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

