% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pseudotimevisualize.R
\name{pseudotimevisualize}
\alias{pseudotimevisualize}
\title{pseudotimevisualize}
\usage{
pseudotimevisualize(pattern, gene, showtrue = F)
}
\arguments{
\item{pattern}{The exact output of the pseudotimepattern function.}

\item{gene}{A character value or vector of gene names. Should be included in the gene expression matrix.}

\item{showtrue}{For the heatmap of multiple gene, whether to display true gene expression values or fitted gene expression values.}
}
\value{
A ggplot2 object
}
\description{
Visualize Gene Expression Pattern for Pseudo Temporal Cell Ordering
}
\details{
Visualize gene expression pattern of one or multiple genes for pseudo temporal cell ordering. For one gene, a scatterplot with fitted lines will be generated.
For multiple genes, a heatmap with fitted or true expression values will be generated.
}
\examples{
data(HSMMdata)
pattern <- pseudotimepattern(HSMMdata,pseudotime)
pseudotimevisualize(pattern,"ENSG00000108821.9")
pseudotimevisualize(pattern,c("ENSG00000108821.9","ENSG00000187193.8"))
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

