% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa-package.R
\docType{package}
\name{sepa-package}
\alias{sepa-package}
\title{sepa: Retrieve data from the Scottish Environmental Protection Agency}
\description{
Functions to discover and download data from the Scottish Environmental Protection Agency (SEPA) API.
}
\examples{
\dontrun{
print("Hello, world")
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simonmoulds/sepa}
}

}
\author{
\strong{Maintainer}: Simon Moulds \email{simon.moulds@gmail.com} (\href{https://orcid.org/0000-0002-7297-482X}{ORCID})

}
