% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEPA.R
\docType{package}
\name{SEPA}
\alias{SEPA}
\alias{SEPA-package}
\title{SEPA: Single-Cell Gene Expression Pattern Analysis}
\description{
SEPA provides useful functions to analysis gene expression patterns and perform GO analysis for single-cell RNA-seq data.
}
\details{
Given single-cell RNA-seq data and true experiment time of cells or pseudo-time cell ordering, SEPA provides convenient functions
for users to assign genes into different gene expression patterns such as constant, monotone increasing and increasing then decreasing.
SEPA then performs GO enrichment analysis to analysis the functional roles of genes with same or similar patterns.
SEPA comes with a user-friendly Graphical User Interface written in shiny.
}

