% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_get_data}
\alias{sepa_get_data}
\alias{sepa_daily_flow}
\alias{sepa_daily_level}
\alias{sepa_daily_precip}
\alias{sepa_15min_flow}
\alias{sepa_15min_level}
\alias{sepa_15min_precip}
\alias{sepa_15min_flow_realtime}
\title{Get SEPA timeseries data}
\usage{
sepa_get_data(ts_id, from, to, ...)

sepa_daily_flow(station_id, from = NULL, to = NULL, ...)

sepa_daily_level(station_id, from = NULL, to = NULL, ...)

sepa_daily_precip(station_id, from = NULL, to = NULL, ...)

sepa_15min_flow(station_id, from = NULL, to = NULL, ...)

sepa_15min_level(station_id, from = NULL, to = NULL, ...)

sepa_15min_precip(station_id, from = NULL, to = NULL, ...)

sepa_15min_flow_realtime(station_id = NULL, ...)
}
\arguments{
\item{ts_id}{Character.}

\item{from}{Date or POSIXt.}

\item{to}{Date or POSIXt.}

\item{...}{Additional arguments. None implemented.}

\item{station_id}{Character.}
}
\value{
A tibble.
}
\description{
Get SEPA timeseries data
}
