% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_get_parameters}
\alias{sepa_get_parameters}
\title{Get available parameters for a set of SEPA sites}
\usage{
sepa_get_parameters(id = NULL, ...)
}
\arguments{
\item{id}{Character. If provided, only information for this
site will be returned. Otherwise information for all sites
is returned.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble containing one row for each station ID and
parameter combination.
}
\description{
Get available parameters for a set of SEPA sites
}
