% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_get_sites}
\alias{sepa_get_sites}
\alias{sepa_get_flow_sites}
\alias{sepa_get_level_sites}
\alias{sepa_get_precip_sites}
\title{Get sites from SEPA}
\usage{
sepa_get_sites(param = NULL, ...)

sepa_get_flow_sites(...)

sepa_get_level_sites(...)

sepa_get_precip_sites(...)
}
\arguments{
\item{param}{Character. If provided, only sites measuring this
parameter will be returned. Otherwise all sites will be return.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
This function gets the table of hydrological stations
}
