% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_get_ts}
\alias{sepa_get_ts}
\title{Get available time series data for a parameter and site}
\usage{
sepa_get_ts(id, param, ...)
}
\arguments{
\item{id}{Character. Station ID.}

\item{param}{Character. Parameter name.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Get available time series data for a parameter and site
}
