% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_group_list.R
\name{sepa_group_list}
\alias{sepa_group_list}
\title{Get available groups}
\usage{
sepa_group_list()
}
\value{
A tibble.
}
\description{
Returns information about the available measurement stations groups.
Group names include "StationsWithFlow", "StationsWithRainfall", amongst
others. Group IDs can be used to filter other queries such as
\code{\link[=sepa_station_list]{sepa_station_list()}}.
}
\examples{
\dontrun{
grps <- sepa_group_list()
}
}
