% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_station_list}
\alias{sepa_station_list}
\title{Get tibble with station information}
\usage{
sepa_station_list(
  station_search_term,
  bounding_box,
  group_id,
  return_fields,
  ...
)
}
\arguments{
\item{station_search_term}{Character. A string pattern to search
station names. This argument supports wildcard matching
using the asterisk (*) symbol. The search is case
insensitive.}

\item{bounding_box}{Numeric. A bounding box within which
to search for stations. This argument should be a vector
with exactly four values formatted as follows:
(min_x, min_y, max_x, max_y).}

\item{group_id}{Character or numeric. A station group id (see sepa_group_list)}

\item{return_fields}{Character. A vector of return fields.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Returns all available stations in the SEPA network
}
