% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SEPAui.R
\name{SEPAui}
\alias{SEPAui}
\title{SEPAui}
\usage{
SEPAui()
}
\description{
Launch the SEPA user interface in local machine
}
\details{
This function will automatically launch the SEPA user interface in a web browser.
It provides a much easier and more convenient way to analysis gene expression patterns and perform GO analysis for single-cell RNA-seq data.
The user interface can also be accessed by http://zhiji.shinyapps.io/SEPA. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
\dontrun{
   SEPAui()
}
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

