% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truetime.R
\name{truetime}
\alias{truetime}
\title{True Experiment Time Points for HSMM Sinlge-cell RNA-seq data}
\format{A data.frame with two columns}
\source{
\url{http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE52529}
}
\description{
The dataset contains a data.frame with two columns specifying true experiment time points. Note that only cells on the main path are retained. Cells belonging to the contaminated path are excluded. Please check HSMMSingleCell package for details.
}
\references{
Trapnell, C., Cacchiarelli, D., Grimsby, J., Pokharel, P., Li, S., Morse, M., ... & Rinn, J. L. (2014). The dynamics and regulators of cell fate decisions are revealed by pseudotemporal ordering of single cells. Nature biotechnology.
}

