% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truetimepattern.R
\name{truetimepattern}
\alias{truetimepattern}
\title{truetimepattern}
\usage{
truetimepattern(expr, truetime, simplify = T, removeconstant = F)
}
\arguments{
\item{expr}{The matrix of gene expression profile.}

\item{truetime}{A character data.frame or matrix of true experimental time. First column: Cell name; Second column: experiment time.}

\item{simplify}{Whether to simplify pattern so that same neiboring patterns will be reduced to one. For example "up_up_constant" will be simplied to "up_constant".}

\item{removeconstant}{Whether to remove all constant patterns. For example "up_up_constant" will be simplied to "up_up". This step will be performed before simplify.}
}
\value{
A named vector or patterns. The names corresponds to gene names.
}
\description{
Identify Pattern for True Experiment Time
}
\details{
Identify the gene expression patterns for true experiment time. For the expressions of each gene, the function performs t-tests for cells from neighboring time points.
The expression pattern for cells from neiboring time points could be increasing, decreasing or constant. All patterns are concatenated using "-" to form the final pattern.
}
\examples{
data(HSMMdata)
truetimepattern(HSMMdata,truetime)
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

