% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/truetimevisualize.R
\name{truetimevisualize}
\alias{truetimevisualize}
\title{truetimevisualize}
\usage{
truetimevisualize(expr, truetime, gene, mode = c("mean", "median"))
}
\arguments{
\item{expr}{The matrix of gene expression profile.}

\item{truetime}{A character data.frame or matrix of true experimental time. First column: Cell name; Second column: experiment time.}

\item{gene}{A vector of gene names to be plotted.}

\item{mode}{A character value specifying mean or median to be displayed}
}
\value{
A ggplot2 object.
}
\description{
Visualize Gene Expression Pattern for True Experiment Time
}
\details{
Identify the gene expression patterns for true experiment time. For the expressions of each gene, the function performs t-tests for cells from neighboring time points.
The expression pattern for cells from neiboring time points could be increasing, decreasing or constant. All patterns are concatenated using "-" to form the final pattern.
}
\examples{
data(HSMMdata)
truetimevisualize(HSMMdata,truetime,c("ENSG00000122180.4","ENSG00000125968.7"))
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

