% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/windowGOanalysis.R
\name{windowGOanalysis}
\alias{windowGOanalysis}
\title{windowGOanalysis}
\usage{
windowGOanalysis(pattern, type = "constant_up", windowsize = NULL,
  movesize = NULL, termnum = 20, identifier = "ENSEMBL",
  species = "Human")
}
\arguments{
\item{pattern}{The direct output of pseudotimepattern function.}

\item{type}{The type of transition pattern.}

\item{windowsize}{The number of genes in each group.}

\item{movesize}{How many genes to move forward each time.}

\item{termnum}{Number of top GO terms to be displayed.}

\item{identifier}{The identifier of the genes. It should be one of the following: "Entrez", "GenBank", "Alias", "Ensembl", "Gene", "Symbol", "GeneName" and "UniGene"}

\item{species}{The species of the genes. Currently only "Human" and "Mouse" are supported}
}
\value{
A list where each element is a data.frame containing the results of GO analysis. The name of the list specifies the group of genes.
}
\description{
Performs GO analysis with moving window for transition patterns
}
\details{
This function is specifically designed for GO analysis of genes with a specific transition pattern. GO analyses are performed iteratively on a group of genes
with similar transition points. Users can define the windowsize (the number of genes in each group) and the movesize (how many genes to move forward each time).
}
\examples{
library(topGO)
data(HSMMdata)
pattern <- pseudotimepattern(HSMMdata,pseudotime)
windowGOanalysis(pattern,type="constant_up")
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

