% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/windowGOvisualize.R
\name{windowGOvisualize}
\alias{windowGOvisualize}
\title{windowGOvisualize}
\usage{
windowGOvisualize(GOres, GOTerm = NULL, topterm = 2, mode = "Heatmap")
}
\arguments{
\item{GOres}{The direct output of windowGOanalysis function.}

\item{GOTerm}{The name of GO term to be displayed. If NULL, top GO terms will be displayed instead}

\item{topterm}{The number of top GO terms to be displayed. This argument only works when GOTerm is NULL.}

\item{mode}{To plot in heatmap or line graph. Either "Heatmap" or "Line".}
}
\value{
A ggplot2 object.
}
\description{
Visualize results of GO analysis with moving window for transition patterns
}
\details{
This function is specifically designed to visualize the results obtained from windowGOanalysis function. Users can choose to visualize specific GO terms or
top GO terms in each time window
}
\examples{
library(topGO)
data(HSMMdata)
pattern <- pseudotimepattern(HSMMdata,pseudotime)
windowGOvisualize(windowGOanalysis(pattern,type="constant_up"))
}
\author{
Zhicheng Ji, Hongkai Ji <zji4@zji4.edu>
}

