% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sepa.R
\name{print.sepa_result}
\alias{print.sepa_result}
\title{Print method for sepa_result objects}
\usage{
\method{print}{sepa_result}(x, ...)
}
\arguments{
\item{x}{A \code{sepa_result} object.}

\item{...}{Ignored.}
}
\description{
Print method for sepa_result objects
}
