% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_get_sites}
\alias{sepa_get_sites}
\title{Get sites from SEPA}
\usage{
sepa_get_sites(...)
}
\arguments{
\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
This function gets the table of hydrological stations
}
