% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_group_list}
\alias{sepa_group_list}
\title{Get tibble with available groups}
\usage{
sepa_group_list()
}
\value{
A tibble.
}
\description{
Returns all available groups. This can be used to filter
other queries such as \code{sepa_station_list}.
}
