% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_timeseries_list.R
\name{sepa_timeseries_list}
\alias{sepa_timeseries_list}
\title{Get list of available time series for one or more stations}
\usage{
sepa_timeseries_list(
  station_id,
  ts_name,
  coverage = TRUE,
  group_id,
  return_fields,
  ...
)
}
\arguments{
\item{station_id}{Character or numeric. Either a single station ID or a
vector of station IDs.}

\item{ts_name}{Character. A time series short name to search for.  This
argument supports wildcard matching using the asterisk (*) symbol.}

\item{coverage}{Logical. Whether or not to return the start and end dates
of the time series. Default is TRUE, but setting to FALSE will result
in faster queries.}

\item{group_id}{Character or numeric. A station group id. See
\code{\link[=sepa_group_list]{sepa_group_list()}}.}

\item{return_fields}{Character. A vector of return fields. In
most cases the default return fields will suffice. See details.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Get list of available time series for one or more stations
}
