% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepa_get_site_list.R
\name{sepa_timeseries_values}
\alias{sepa_timeseries_values}
\title{Get list of available time series for one or more stations}
\usage{
sepa_timeseries_values(ts_id, start_date, end_date, return_fields, ...)
}
\arguments{
\item{ts_id}{Character or numeric. Either a single station ID or a
vector of station IDs. Time series IDs can be found using the
\code{sepa_timeseries_list} function.}

\item{start_date}{Date or character formatted "YYYY-MM-DD". Defaults to yesterday.}

\item{end_date}{Date or character formatted "YYYY-MM-DD". Defaults to today.}

\item{return_fields}{Character. A vector of return fields.}

\item{...}{Additional arguments. None implemented.}
}
\value{
A tibble.
}
\description{
Get list of available time series for one or more stations
}
